Preserve union order in schema representations
