+++ packages/effect/src/internal/schema/schema.ts
@@ -91,45 +91,3 @@ export function fromIssueEffect<A, R>(
     (cause) => Effect.failCauseSync(() => Cause.map(cause, (issue) => new SchemaError(issue)))
   )
 }
-
-/** @internal */
-export const jsonReorder = makeReorder(getJsonPriority)
-
-function getJsonPriority(ast: SchemaAST.AST): number {
-  switch (ast._tag) {
-    case "BigInt":
-    case "Symbol":
-    case "UniqueSymbol":
-      return 0
-    default:
-      return 1
-  }
-}
-
-/** @internal */
-export function makeReorder(getPriority: (ast: SchemaAST.AST) => number) {
-  return (types: ReadonlyArray<SchemaAST.AST>): ReadonlyArray<SchemaAST.AST> => {
-    // Create a map of original indices for O(1) lookup
-    const indexMap = new Map<SchemaAST.AST, number>()
-    for (let i = 0; i < types.length; i++) {
-      indexMap.set(SchemaAST.toEncoded(types[i]), i)
-    }
-
-    // Create a sorted copy of the types array
-    const sortedTypes = [...types].sort((a, b) => {
-      a = SchemaAST.toEncoded(a)
-      b = SchemaAST.toEncoded(b)
-      const pa = getPriority(a)
-      const pb = getPriority(b)
-      if (pa !== pb) return pa - pb
-      // If priorities are equal, maintain original order (stable sort)
-      return indexMap.get(a)! - indexMap.get(b)!
-    })
-
-    // Check if order changed by comparing arrays
-    const orderChanged = sortedTypes.some((ast, index) => ast !== types[index])
-
-    if (!orderChanged) return types
-    return sortedTypes
-  }
-}
